Split AtomicTest away from PrimitivesTest
