+++ lincheck/src/jvm/test/org/jetbrains/kotlinx/lincheck_test/strategy/eventstructure/PrimitivesTest.kt
@@ -1648,166 +1648,6 @@ class PrimitivesTest {
         }
     }
 
-    @Test
-    fun testThreeFAA() {
-        val outcomes = setOf<List<Int>>(
-            listOf(0,1,2),
-            listOf(0,2,1),
-            listOf(1,0,2),
-            listOf(1,2,0),
-            listOf(2,0,1),
-            listOf(2,1,0),
-        )
-        litmusTest(assertSame(outcomes)) {
-            val x = AtomicInteger(0)
-            var r0 = -1
-            var r1 = -1
-            var r2 = -1
-
-            val t0 = thread {
-                r0 = x.getAndIncrement()
-            }
-            val t1 = thread {
-                r1 = x.getAndIncrement()
-            }
-            val t2 = thread {
-                r2 = x.getAndIncrement()
-            }
-
-            t0.join()
-            t1.join()
-            t2.join()
-
-            listOf(r0, r1, r2)
-        }
-    }
-
-    @Test
-    fun testRMW3_1() {
-        val outcomes = setOf<List<Int>>(
-            listOf(0,1,2,3),
-            listOf(0,3,1,2),
-            listOf(0,2,1,3),
-            listOf(2,3,0,1),
-            listOf(1,2,0,3),
-            listOf(1,3,0,2),
-        )
-        litmusTest(assertSame(outcomes)) {
-            val x = AtomicInteger(0)
-
-            val r = IntArray(4)
-
-            val t1 = thread {
-                r[0] = x.getAndIncrement()
-                r[1] = x.getAndIncrement()
-            }
-            val t2 = thread {
-                r[2] = x.getAndIncrement()
-                r[3] = x.getAndIncrement()
-            }
-
-            t1.join()
-            t2.join()
-
-            listOf(r[0], r[1], r[2], r[3])
-        }
-    }
-
-    @Test
-    fun testRMW3_2() {
-        val outcomes = setOf<List<Int>>(
-            listOf(0,1,2,3),
-            listOf(0,3,1,2),
-            listOf(0,2,1,3),
-            listOf(2,3,0,1),
-            listOf(1,2,0,3),
-            listOf(1,3,0,2),
-        )
-        class TestClass {
-            val x = AtomicInteger(0)
-            fun one(): Pair<Int, Int> {
-                val r1 = x.getAndIncrement()
-                val r2 = x.getAndIncrement()
-                return r1 to r2
-            }
-
-            fun two(): Pair<Int, Int> {
-                val r3 = x.getAndIncrement()
-                val r4 = x.getAndIncrement()
-                return r3 to r4
-            }
-        }
-
-        val scenario = scenario {
-            parallel {
-                thread { actor(TestClass::one) }
-                thread { actor(TestClass::two) }
-            }
-        }
-
-        litmusTest(TestClass::class.java, scenario, assertSame(outcomes)) { results ->
-            val p1 = getValue<Pair<Int, Int>>(results.parallelResults[0][0]!!)
-            val p2 = getValue<Pair<Int, Int>>(results.parallelResults[1][0]!!)
-
-            listOf(p1.first, p1.second, p2.first, p2.second)
-        }
-    }
-
-
-    @Test
-    fun testCAS() {
-        val outcomes = setOf<List<Int>>(
-            listOf(1,0,0),
-            listOf(0,1,0),
-            listOf(0,0,1),
-        )
-        litmusTest(assertSame(outcomes)) {
-            val x = AtomicInteger(0)
-            val r = IntArray(3)
-
-            val t1 = thread {
-                r[0] = if (x.compareAndSet(0, 1)) 1 else 0
-            }
-            val t2 = thread {
-                r[1] = if (x.compareAndSet(0, 1)) 1 else 0
-            }
-            val t3 = thread {
-                r[2] = if (x.compareAndSet(0, 1)) 1 else 0
-            }
-
-            t1.join()
-            t2.join()
-            t3.join()
-
-            listOf(r[0], r[1], r[2])
-        }
-    }
-
-    @Test
-    fun testCAS2() {
-        val outcomes = setOf<List<Int>>(
-            listOf(1,0),
-            listOf(0,1)
-        )
-        litmusTest(assertSame(outcomes)) {
-            val x = AtomicInteger(0)
-            val r = IntArray(2)
-
-            val t1 = thread {
-                r[0] = if (x.weakCompareAndSetPlain(0, 3)) 1 else 0
-                if (r[0] == 0) r[1] = x.getAndIncrement()
-            }
-
-            val t2 = thread {
-                x.setOpaque(1)
-            }
-
-            t1.join()
-            t2.join()
-
-            listOf(r[0], r[1])
-        }
-    }
 }
